refactor(webhook): move base64-to-file utility to pieces-common, use file-type for full MIME detection
